Move OpenApiConfigMapping to runtime
This SPI implementation was erroneously part of the
deployment module instead of the runtime module
