--- extensions/smallrye-openapi/deployment/src/main/resources/META-INF/services/io.smallrye.config.ConfigSourceInterceptor
@@ -1,1 +1,0 @@
-io.quarkus.smallrye.openapi.deployment.OpenApiConfigMapping
