+++ extensions/smallrye-openapi/runtime/src/main/resources/META-INF/services/io.smallrye.config.ConfigSourceInterceptor
@@ -1,0 +1,1 @@
+io.quarkus.smallrye.openapi.runtime.OpenApiConfigMapping
